compiled nut_therm_shell test for better efficiency, along with EOSsuper built into sedonu

+++ src/grid_1D_sphere.cpp
@@ -90,7 +90,7 @@ int grid_1D_sphere::zone_index(const vector<double>& x) const
 	// find in zone array using stl algorithm upper_bound and subtracting iterators
 	int z_ind = r_out.locate(r);
 	assert(z_ind >= 0);
-	if(z_ind>=z.size()){
+	if(z_ind>=(int)z.size()){
 		cout << z_ind << endl;
 		cout << z.size() << endl;
 		cout << r << endl;
